pyproject.toml
[project]
name = "pricehawk"
version = "0.1.0"
description = "Monitor competitor prices, detect changes, and get AI-powered insights"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "supabase>=2.27.0",
    "uvicorn[standard]>=0.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "playwright>=1.40.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
]
